cstream-3.1.1.tar.gz tarball content

+++ cstream.c
@@ -1678,7 +1678,7 @@ atoi_kmg(const char *const s)
   long long res;
   char c;
 
-  res = atoi(s);
+  res = atoll(s);
   if (s[0] != '\0') {
     c = tolower(s[strlen(s)-1]);
     switch (c) {
